feat: recommend - user 유저엔티티 regionId 추가, 컨트롤러 새로운 버전 추가 (주석은 모든 테스트 후 제거예정), 게이트웨이 api주소 추가, postgis geometry설정 추가
