+++ user/src/main/java/com/team15gijo/user/application/config/GeometryConfig.java
@@ -1,0 +1,15 @@
+package com.team15gijo.user.application.config;
+
+import org.locationtech.jts.geom.GeometryFactory;
+import org.locationtech.jts.geom.PrecisionModel;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class GeometryConfig {
+
+    @Bean
+    public GeometryFactory geometryFactory() {
+        return new GeometryFactory(new PrecisionModel(), 4326);
+    }
+}
